chore: rename example, track in root toml, and update lock file
