+++ docs/source/_static/hybrid_control_flow.png
